requirements.txt
Flask
requests
flask-caching
google-generativeai
python-dotenv